requirements.txt
aggdraw==1.3.12
attrs==21.2.0
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
psd-tools==1.9.17
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
tifffile==2021.7.2
tornado==6.1
xlrd==2.0.1
xlwt==1.3.0
